@@ -45,13 +45,14 @@ LocalConference::LocalConference (
 	const shared_ptr<Core> &core,
 	const IdentityAddress &myAddress,
 	CallSessionListener *listener,
-	const std::shared_ptr<ConferenceParams> params
+	const std::shared_ptr<ConferenceParams> params ,
+	ConferenceListener *confListener
 	) : Conference(core, myAddress, listener, params) {
 	// Set last notify to 1 in order to ensure that the 1st notify to remote conference is correctly processed
 	// Remote conference sets last notify to 0 in its constructor
 	lastNotify = 1;
 #ifdef HAVE_ADVANCED_IM
-	eventHandler = std::make_shared<LocalConferenceEventHandler>(this);
+	eventHandler = std::make_shared<LocalConferenceEventHandler>(this, confListener);
 	addListener(eventHandler);
 #endif
 }
