@@ -650,7 +650,7 @@ void first_notify_parsing() {
 	size_t size = strlen(first_notify) + strlen(confUri);
 	char *notify = new char[size];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
@@ -688,7 +688,7 @@ void first_notify_parsing_wrong_conf() {
 	size_t size = strlen(first_notify) + strlen(confUri);
 	char *notify = new char[size];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
 
@@ -720,7 +720,7 @@ void participant_added_parsing() {
 	size_t size2 = strlen(participant_added_notify) + strlen(confUri);
 	char *notify_added = new char[size2];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
 
@@ -766,7 +766,7 @@ void participant_not_added_parsing() {
 	size_t size2 = strlen(participant_not_added_notify) + strlen(confUri);
 	char *notify_not_added = new char[size2];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
 
@@ -809,7 +809,7 @@ void participant_deleted_parsing() {
 	size_t size2 = strlen(participant_deleted_notify) + strlen(confUri);
 	char *notify_deleted = new char[size2];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
 
@@ -852,7 +852,7 @@ void participant_admined_parsing() {
 	size_t size2 = strlen(participant_admined_notify) + strlen(confUri);
 	char *notify_admined = new char[size2];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
 
@@ -894,7 +894,7 @@ void participant_unadmined_parsing() {
 	size_t size2 = strlen(participant_unadmined_notify) + strlen(confUri);
 	char *notify_unadmined = new char[size2];
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	snprintf(notify, size, first_notify, confUri);
 	tester->handler->notifyReceived(notify);
 
@@ -952,7 +952,7 @@ void send_first_notify() {
 	localConf->setConferenceAddress(ConferenceAddress(addr));
 	string notify = localHandler->createNotifyFullState();
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	tester->handler->notifyReceived(notify);
 
 	BC_ASSERT_STRING_EQUAL(tester->confSubject.c_str(), "A random test subject");
@@ -1699,7 +1699,7 @@ void one_to_one_keyword () {
 	localConf->setConferenceAddress(ConferenceAddress(addr));
 	string notify = localHandler->createNotifyFullState(true);
 
-	const_cast<IdentityAddress &>(tester->handler->getConferenceId().getPeerAddress()) = addr;
+	const_cast<ConferenceAddress &>(tester->handler->getConferenceId().getPeerAddress()) = ConferenceAddress(addr);
 	tester->handler->notifyReceived(notify);
 
 	BC_ASSERT_EQUAL(tester->participantDevices.size(), 1, int, "%d");
